+++ NEWS.md
@@ -6,6 +6,7 @@
    
 - New function `theme_classic2()` added. Classic theme with axis lines.
 - New argument star.plot in ggscatter(). A logical value. If TRUE, a star plot is generated.
+- New helper function `geom_exec()`. A helper function used by ggpubr functions to execute any geom_xx functions in ggplot2. Useful only when you want to call a geom_xx function without carrying about the arguments to put in `ggplot2::aes()`.
 
 ## Minor changes